+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## 0.4.0
+ - added `-q` option to allow query for contents of certain DOM elements via CSS selector;
+ - constructors of the `Reader` are now returning reference instead of value.
+
 ## 0.3.1
  - fixed bug in `Reader#Read` leading to stack overflow error.